+++ app/controllers/ControllerBase.php
@@ -1,14 +1,30 @@
 <?php
+
+/*
+  +------------------------------------------------------------------------+
+  | Vökuró                                                                 |
+  +------------------------------------------------------------------------+
+  | Copyright (c) 2016-present Phalcon Team (https://www.phalconphp.com)   |
+  +------------------------------------------------------------------------+
+  | This source file is subject to the New BSD License that is bundled     |
+  | with this package in the file LICENSE.txt.                             |
+  |                                                                        |
+  | If you did not receive a copy of the license and are unable to         |
+  | obtain it through the world-wide-web, please send an email             |
+  | to license@phalconphp.com so we can send you a copy immediately.       |
+  +------------------------------------------------------------------------+
+*/
+
 namespace Vokuro\Controllers;
 
 use Phalcon\Mvc\Controller;
 use Phalcon\Mvc\Dispatcher;
 
 /**
- * ControllerBase
- * This is the base controller for all controllers in the application
- *
+ * ControllerBase. This is the base controller for all controllers in the application
+ * Vokuro\Controllers\ProfilesController
  * @property \Vokuro\Auth\Auth auth
+ * @package Vokuro\Controllers
  */
 class ControllerBase extends Controller
 {
@@ -25,13 +41,11 @@ class ControllerBase extends Controller
 
         // Only check permissions on private controllers
         if ($this->acl->isPrivate($controllerName)) {
-
             // Get the current identity
             $identity = $this->auth->getIdentity();
 
             // If there is no identity available the user is redirected to index/index
             if (!is_array($identity)) {
-
                 $this->flash->notice('You don\'t have access to this module: private');
 
                 $dispatcher->forward([
@@ -44,7 +58,6 @@ class ControllerBase extends Controller
             // Check if the user have permission to the current option
             $actionName = $dispatcher->getActionName();
             if (!$this->acl->isAllowed($identity['profile'], $controllerName, $actionName)) {
-
                 $this->flash->notice('You don\'t have access to this module: ' . $controllerName . ':' . $actionName);
 
                 if ($this->acl->isAllowed($identity['profile'], $controllerName, 'index')) {