+++ app/controllers/PermissionsController.php
@@ -1,4 +1,20 @@
 <?php
+
+/*
+  +------------------------------------------------------------------------+
+  | Vökuró                                                                 |
+  +------------------------------------------------------------------------+
+  | Copyright (c) 2016-present Phalcon Team (https://www.phalconphp.com)   |
+  +------------------------------------------------------------------------+
+  | This source file is subject to the New BSD License that is bundled     |
+  | with this package in the file LICENSE.txt.                             |
+  |                                                                        |
+  | If you did not receive a copy of the license and are unable to         |
+  | obtain it through the world-wide-web, please send an email             |
+  | to license@phalconphp.com so we can send you a copy immediately.       |
+  +------------------------------------------------------------------------+
+*/
+
 namespace Vokuro\Controllers;
 
 use Vokuro\Models\Profiles;
@@ -6,10 +22,11 @@ use Vokuro\Models\Permissions;
 
 /**
  * View and define permissions for the various profile levels.
+ * Vokuro\Controllers\PermissionsController
+ * @package Vokuro\Controllers
  */
 class PermissionsController extends ControllerBase
 {
-
     /**
      * View the permissions for a profile level, and change them if we have a POST.
      */
@@ -18,20 +35,16 @@ class PermissionsController extends ControllerBase
         $this->view->setTemplateBefore('private');
 
         if ($this->request->isPost()) {
-
             // Validate the profile
             $profile = Profiles::findFirstById($this->request->getPost('profileId'));
 
             if ($profile) {
-
                 if ($this->request->hasPost('permissions') && $this->request->hasPost('submit')) {
-
                     // Deletes the current permissions
                     $profile->getPermissions()->delete();
 
                     // Save the new permissions
                     foreach ($this->request->getPost('permissions') as $permission) {
-
                         $parts = explode('.', $permission);
 
                         $permission = new Permissions();