+++ app/forms/ChangePasswordForm.php
@@ -1,4 +1,20 @@
 <?php
+
+/*
+  +------------------------------------------------------------------------+
+  | Vökuró                                                                 |
+  +------------------------------------------------------------------------+
+  | Copyright (c) 2016-present Phalcon Team (https://www.phalconphp.com)   |
+  +------------------------------------------------------------------------+
+  | This source file is subject to the New BSD License that is bundled     |
+  | with this package in the file LICENSE.txt.                             |
+  |                                                                        |
+  | If you did not receive a copy of the license and are unable to         |
+  | obtain it through the world-wide-web, please send an email             |
+  | to license@phalconphp.com so we can send you a copy immediately.       |
+  +------------------------------------------------------------------------+
+*/
+
 namespace Vokuro\Forms;
 
 use Phalcon\Forms\Form;
@@ -7,12 +23,14 @@ use Phalcon\Validation\Validator\PresenceOf;
 use Phalcon\Validation\Validator\StringLength;
 use Phalcon\Validation\Validator\Confirmation;
 
+/**
+ * Vokuro\Forms\ChangePasswordForm
+ * @package Vokuro\Forms
+ */
 class ChangePasswordForm extends Form
 {
-
     public function initialize()
     {
-        // Password
         $password = new Password('password');
 
         $password->addValidators([