+++ app/library/Auth/Exception.php
@@ -1,6 +1,0 @@
-<?php
-namespace Vokuro\Auth;
-
-class Exception extends \Exception
-{
-}