+++ app/models/EmailConfirmations.php
@@ -1,39 +1,45 @@
 <?php
+
+/*
+  +------------------------------------------------------------------------+
+  | Vökuró                                                                 |
+  +------------------------------------------------------------------------+
+  | Copyright (c) 2016-present Phalcon Team (https://www.phalconphp.com)   |
+  +------------------------------------------------------------------------+
+  | This source file is subject to the New BSD License that is bundled     |
+  | with this package in the file LICENSE.txt.                             |
+  |                                                                        |
+  | If you did not receive a copy of the license and are unable to         |
+  | obtain it through the world-wide-web, please send an email             |
+  | to license@phalconphp.com so we can send you a copy immediately.       |
+  +------------------------------------------------------------------------+
+*/
+
 namespace Vokuro\Models;
 
 use Phalcon\Mvc\Model;
 
 /**
- * EmailConfirmations
- * Stores the reset password codes and their evolution
+ * EmailConfirmations. Stores the reset password codes and their evolution
+ * Vokuro\Models\EmailConfirmations
+ * @method static EmailConfirmations findFirstByCode($code)
+ * @package Vokuro\Models
  */
 class EmailConfirmations extends Model
 {
-
-    /**
-     *
-     * @var integer
-     */
+    /** @var integer */
     public $id;
 
-    /**
-     *
-     * @var integer
-     */
+    /** @var integer */
     public $usersId;
 
+    /** @var string */
     public $code;
 
-    /**
-     *
-     * @var integer
-     */
+    /** @var integer */
     public $createdAt;
 
-    /**
-     *
-     * @var integer
-     */
+    /** @var integer */
     public $modifiedAt;
 
     public $confirmed;