+++ app/models/Profiles.php
@@ -1,26 +1,37 @@
 <?php
+
+/*
+  +------------------------------------------------------------------------+
+  | Vökuró                                                                 |
+  +------------------------------------------------------------------------+
+  | Copyright (c) 2016-present Phalcon Team (https://www.phalconphp.com)   |
+  +------------------------------------------------------------------------+
+  | This source file is subject to the New BSD License that is bundled     |
+  | with this package in the file LICENSE.txt.                             |
+  |                                                                        |
+  | If you did not receive a copy of the license and are unable to         |
+  | obtain it through the world-wide-web, please send an email             |
+  | to license@phalconphp.com so we can send you a copy immediately.       |
+  +------------------------------------------------------------------------+
+*/
+
 namespace Vokuro\Models;
 
 use Phalcon\Mvc\Model;
 use Phalcon\Mvc\Model\Relation;
 
 /**
- * Vokuro\Models\Profiles
  * All the profile levels in the application. Used in conjenction with ACL lists
+ * Vokuro\Models\Profiles
+ * @method static Profiles findFirstById($id)
+ * @package Vokuro\Models
  */
 class Profiles extends Model
 {
-
-    /**
-     * ID
-     * @var integer
-     */
+    /** @var integer */
     public $id;
 
-    /**
-     * Name
-     * @var string
-     */
+    /** @var string */
     public $name;
 
     /**